refactor(ios-keyvox-speak): unify the intro and unlock sheets into one shared screen
- add `iOS/KeyVox iOS/Views/Components/KeyVoxSpeakSheetView.swift` as the shared KeyVox Speak presentation surface used by both the post-onboarding intro flow and the unlock flow
- move the shared three-page swipeable KeyVox Speak content, page dots, pinned bottom action area, unlock summary copy, and unlock purchase or restore affordances into `KeyVoxSpeakSheetView.swift` so the intro and monetization paths no longer duplicate the same information
- update `iOS/KeyVox iOS/Views/Components/KeyVoxSpeakIntroSheetView.swift` to become a thin wrapper around `KeyVoxSpeakSheetView` in intro mode with the single `Try KeyVox Speak` call to action
- update `iOS/KeyVox iOS/Views/Components/TTSUnlockSheetView.swift` to become a thin wrapper around `KeyVoxSpeakSheetView` in unlock mode with one large unlock button, a small white `Restore Purchases` action underneath it, and the existing dismiss behavior
- keep the DEBUG scheme force flag enabled in `iOS/KeyVox iOS.xcodeproj/xcshareddata/xcschemes/KeyVox iOS DEBUG.xcscheme` so the unified KeyVox Speak sheet can be previewed repeatedly during this polish pass
