+++ iOS/KeyVox iOS.xcodeproj/xcshareddata/xcschemes/KeyVox iOS DEBUG.xcscheme
@@ -81,7 +81,7 @@
          <EnvironmentVariable
             key = "KEYVOX_FORCE_KEYVOX_SPEAK_INTRO"
             value = "1"
-            isEnabled = "NO">
+            isEnabled = "YES">
          </EnvironmentVariable>
       </EnvironmentVariables>
    </LaunchAction>
